Add Academy-specific JSONs to version control
From now on, we'll include each of the divergent JSONs in a
protocol-specific subdirectory, while the 104 versions remain in the
root as the "master templates".

When making changes to the 104 JSONs that should be propaganted to the
other versions, make the change first as one commit just for 104, then
another commit that propagates the change to *all* the alternate builds.

This system might get replaced when we implement the static DB. We'll
see.